vim: created .vimrc
I had `~/.config/vim`, but I couldn't make it work... back to `~/.vimrc` guys.

I remembered that I don't actually use vim as my main editor, I utilize neovim
for that, so this configuration will only be used when I am in another person
computer, so because of it I will not modularize this configuration (meaning
that this will be kept in one single file), and also I will not add plugins unless
necessary (like vim-tmux-navigator), with the exception of colorscheme... what?
I like beautiful colors on my terminal editor.